Move CallbackPropertiesBase into lib
